+++ packages/nova-react-test-utils/src/shared/types.ts
@@ -1,5 +1,51 @@
 import type { StoryObj } from "@storybook/react";
-import type { Simplify, RequiredKeysOf } from "type-fest";
+import type { Simplify, RequiredKeysOf, IfAny, SimplifyDeep } from "type-fest";
+import type { ComponentType } from "react";
+
+/**
+ * Utility type to extract the props of a React component.
+ */
+export type Props<Component> =
+  Component extends ComponentType<infer P> ? P : never;
+
+/**
+ * Utility type to return the keys of a component's props, where the value of the key is an object containing an optional " $data" property.
+ * This is typically used to identify GraphQL fragment reference properties.
+ */
+export type FragmentKeys<Component> =
+  Props<Component> extends infer P
+    ? Required<{
+        [K in keyof P]: IfAny<
+          P[K],
+          never,
+          P[K] extends { " $data"?: unknown } ? K : never
+        >;
+      }>[keyof P]
+    : never;
+
+/**
+ * Utility type to get the data type of a fragment reference from a React component.
+ */
+export type FragmentDataType<
+  Component,
+  K extends FragmentKeys<Component>,
+> = Props<Component>[K] extends {
+  " $data"?: infer D extends { " $fragmentType": unknown };
+}
+  ? Omit<D, " $fragmentType">
+  : never;
+
+/**
+ * Utility type to get the fragment key type from a React component.
+ */
+export type FragmentKeyType<
+  Component,
+  K extends FragmentKeys<Component>,
+> = Props<Component>[K] extends {
+  readonly " $fragmentSpreads": unknown;
+}
+  ? Props<Component>[K]
+  : never;
 
 type RequiredArgs<T> = T extends { args?: infer A }
   ? A extends object
@@ -14,13 +60,12 @@ type OptionalArgs<T> = T extends { args?: infer A }
 
 // Omits Z from the Storybook "args" field of T
 type OmitFromArgs<T, Z> = Omit<T, "args"> &
-  Simplify<
+  SimplifyDeep<
     (Record<string, never> extends Omit<RequiredArgs<T>, keyof Z>
       ? { args?: Omit<RequiredArgs<T>, keyof Z> }
       : { args: Omit<RequiredArgs<T>, keyof Z> }) & {
       args?: Omit<OptionalArgs<T>, keyof Z>;
-    },
-    { deep: true }
+    }
   >;
 
 // Check if keys of T are a subset of keys of P. Returns true if they are or union of keys that are not.
@@ -41,11 +86,41 @@ export type ComponentTypeError =
 export type ReferenceEntriesError<K> = `❌ Error: The reference entry '${K &
   string}' is not a property on the component's props`;
 
+/**
+ * Type for validated reference entries when a component type is provided
+ */
+export type ValidatedReferenceEntries<
+  TComponent,
+  TQuery extends { response: unknown },
+> =
+  // The any here is important, changing to unknown or never would break inferring component
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  TComponent extends ComponentType<any>
+    ? FragmentKeys<TComponent> extends never
+      ? Record<string, (queryResult: TQuery["response"]) => unknown>
+      : // Fragment keys must have correct type, other keys can be anything
+        Simplify<
+          {
+            [K in FragmentKeys<TComponent>]: (
+              queryResult: TQuery["response"],
+            ) => FragmentKeyType<TComponent, K> | null | undefined;
+          } & {
+            [K in Exclude<string, FragmentKeys<TComponent>>]?: (
+              queryResult: TQuery["response"],
+            ) => unknown;
+          }
+        >
+    : Record<string, (queryResult: TQuery["response"]) => unknown>;
+
 export type StoryObjWithoutFragmentRefs<T> = T extends {
   component?: infer C;
-  parameters: { novaEnvironment: { referenceEntries: infer D extends object } };
+  parameters: {
+    novaEnvironment: {
+      referenceEntries: infer D extends object;
+    };
+  };
 }
-  ? C extends React.ComponentType<infer P extends object>
+  ? C extends ComponentType<infer P extends object>
     ? AreKeysSubset<D, P> extends true
       ? OmitFromArgs<StoryObj<T>, D>
       : ReferenceEntriesError<AreKeysSubset<D, P>>